jobs/build-node-image: publish extensions first and support multiples tags
The node image being published is the signal for
for ART that nightly content is released.
Publishing the extensions image after can take long
enough that the ART process has already kicked in
and fail because all the expected artifacts are
not present.

Let's make sure we publish the node image as the
very last step.

Also publish a moving tag so we have a "latest"
floating tag.
This gives an easy way to identify what's the most
recent image.

See https://github.com/coreos/fedora-coreos-pipeline/pull/1120#issuecomment-2787686055